+++ src/Telegram.Bot/Types/Payments/RefundedPayment.cs
@@ -11,7 +11,7 @@ public partial class RefundedPayment
     /// <summary>Total refunded price in the <em>smallest units</em> of the currency (integer, <b>not</b> float/double). For example, for a price of <c>US$ 1.45</c>, <c><see cref="TotalAmount">TotalAmount</see> = 145</c>. See the <em>exp</em> parameter in <a href="https://core.telegram.org/bots/payments/currencies.json">currencies.json</a>, it shows the number of digits past the decimal point for each currency (2 for the majority of currencies).</summary>
     [JsonPropertyName("total_amount")]
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public int TotalAmount { get; set; }
+    public long TotalAmount { get; set; }
 
     /// <summary>Bot-specified invoice payload</summary>
     [JsonPropertyName("invoice_payload")]
