+++ src/Telegram.Bot/Types/Payments/StarTransaction.cs
@@ -14,7 +14,7 @@ public partial class StarTransaction
 
     /// <summary><em>Optional</em>. The number of 1/1000000000 shares of Telegram Stars transferred by the transaction; from 0 to 999999999</summary>
     [JsonPropertyName("nanostar_amount")]
-    public int? NanostarAmount { get; set; }
+    public long? NanostarAmount { get; set; }
 
     /// <summary>Date the transaction was created</summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
