+++ src/Telegram.Bot/Types/StarAmount.cs
@@ -10,5 +10,5 @@ public partial class StarAmount
 
     /// <summary><em>Optional</em>. The number of 1/1000000000 shares of Telegram Stars; from -999999999 to 999999999; can be negative if and only if <see cref="Amount">Amount</see> is non-positive</summary>
     [JsonPropertyName("nanostar_amount")]
-    public int? NanostarAmount { get; set; }
+    public long? NanostarAmount { get; set; }
 }
